feat(services): move registry to node-core (#1945)
