+++ mod/node-core/pkg/services/registry/errors.go
