+++ mod/node-core/pkg/services/registry/mocks/basic.mock.go
