+++ mod/node-core/pkg/services/registry/options.go
