+++ mod/node-core/pkg/services/registry/registry.go
